refactor(core): co-locate error-class serializers in errors.ts
Move the RpcError/TypedError registerClassSerializer registrations out of the
standalone src/runtypes/mionClassSerializers.ts and into src/errors.ts, next to the
class definitions they register (loading @mionjs/core still fires them before any
decode, via the errors.ts re-export). Delete mionClassSerializers.ts and relocate its
round-trip spec to src/error-class-serializers.spec.ts (the registration now rides the
errors.ts import). The registration MECHANISM was already ts-runtypes'
(registerClassSerializer); this just removes the separate wrapper module.
